ai-email-processor/requirements.txt
# requirements.txt
asyncpg==0.29.0
openai==1.35.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
schedule==1.2.0
cryptography==41.0.7
email-validator==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1